LayoutTests/svg/dom/SVGPathSegList-xml-dom-synchronization-expected.txt

This is a test how SVGLengthList reacts to XML DOM modifications.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS path.pathSegList.numberOfItems is 3
PASS path.getAttribute('d').formatDAttribute() is "M 200 0 L 100 0 L 100 100"
PASS path.getAttribute('d').formatDAttribute() is "M 200 0 L 100 0 L 100 100 L 0 100"
PASS path.getAttribute('d').formatDAttribute() is "M 100 0 L 100 0 L 100 100 L 0 100"
PASS path.getAttribute('d').formatDAttribute() is "M 0 0 L 100 0 L 100 100 L 0 100"
PASS successfullyParsed is true

TEST COMPLETE
